Migrated retry_failed into a component
