+++ flexget/components/failed/retry_failed.py
@@ -1,87 +1,19 @@
 from __future__ import unicode_literals, division, absolute_import
-from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
 import logging
+from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 from datetime import datetime, timedelta
 
-from sqlalchemy import Column, Integer, String, Unicode, DateTime
-from sqlalchemy.schema import Index
-
-from flexget import db_schema, plugin
+from flexget import plugin
 from flexget.event import event
 from flexget.manager import Session
-from flexget.utils.sqlalchemy_utils import table_add_column
 from flexget.utils.tools import parse_timedelta
+from . import db
 
 SCHEMA_VER = 3
 FAIL_LIMIT = 100
 
 log = logging.getLogger('failed')
-Base = db_schema.versioned_base('failed', SCHEMA_VER)
-
-
-@db_schema.upgrade('failed')
-def upgrade(ver, session):
-    if ver is None or ver < 1:
-        raise db_schema.UpgradeImpossible
-    if ver == 1:
-        table_add_column('failed', 'reason', Unicode, session)
-        ver = 2
-    if ver == 2:
-        table_add_column('failed', 'retry_time', DateTime, session)
-        ver = 3
-    return ver
-
-
-class FailedEntry(Base):
-    __tablename__ = 'failed'
-
-    id = Column(Integer, primary_key=True)
-    title = Column(Unicode)
-    url = Column(String)
-    tof = Column(DateTime)
-    reason = Column(Unicode)
-    count = Column(Integer, default=1)
-    retry_time = Column(DateTime)
-
-    def __init__(self, title, url, reason=None):
-        self.title = title
-        self.url = url
-        self.reason = reason
-        self.tof = datetime.now()
-
-    def __str__(self):
-        return '<Failed(title=%s)>' % self.title
-
-    def to_dict(self):
-        return {
-            'id': self.id,
-            'title': self.title,
-            'url': self.url,
-            'added_at': self.tof,
-            'reason': self.reason,
-            'count': self.count,
-            'retry_time': self.retry_time
-        }
-
-
-# create indexes, used when creating tables
-columns = Base.metadata.tables['failed'].c
-Index('failed_title_url', columns.title, columns.url, columns.count)
-
-
-@event('manager.db_cleanup')
-def db_cleanup(manager, session):
-    # Delete everything older than 30 days
-    session.query(FailedEntry).filter(FailedEntry.tof < datetime.now() - timedelta(days=30)).delete()
-    # Of the remaining, always keep latest 25. Drop any after that if fail was more than a week ago.
-    keep_num = 25
-    keep_ids = [fe.id for fe in session.query(FailedEntry).order_by(FailedEntry.tof.desc())[:keep_num]]
-    if len(keep_ids) == keep_num:
-        query = session.query(FailedEntry)
-        query = query.filter(FailedEntry.id.notin_(keep_ids))
-        query = query.filter(FailedEntry.tof < datetime.now() - timedelta(days=7))
-        query.delete(synchronize_session=False)
 
 
 class PluginFailed(object):
@@ -159,10 +91,10 @@ class PluginFailed(object):
         reason = str(reason) or 'Unknown'
         with Session() as session:
             # query item's existence
-            item = session.query(FailedEntry).filter(FailedEntry.title == entry['title']). \
-                filter(FailedEntry.url == entry['original_url']).first()
+            item = session.query(db.FailedEntry).filter(db.FailedEntry.title == entry['title']). \
+                filter(db.FailedEntry.url == entry['original_url']).first()
             if not item:
-                item = FailedEntry(entry['title'], entry['original_url'], reason)
+                item = db.FailedEntry(entry['title'], entry['original_url'], reason)
                 item.count = 0
             if item.count > FAIL_LIMIT:
                 log.error('entry with title \'%s\' has failed over %s times', entry['title'], FAIL_LIMIT)
@@ -185,8 +117,8 @@ class PluginFailed(object):
         config = self.prepare_config(config)
         max_count = config['max_retries']
         for entry in task.entries:
-            item = task.session.query(FailedEntry).filter(FailedEntry.title == entry['title']). \
-                filter(FailedEntry.url == entry['original_url']).first()
+            item = task.session.query(db.FailedEntry).filter(db.FailedEntry.title == entry['title']). \
+                filter(db.FailedEntry.url == entry['original_url']).first()
             if item:
                 if item.count > max_count:
                     entry.reject('Has already failed %s times in the past. (failure reason: %s)' %
@@ -199,14 +131,3 @@ class PluginFailed(object):
 @event('plugin.register')
 def register_plugin():
     plugin.register(PluginFailed, 'retry_failed', builtin=True, api_ver=2)
-
-
-def get_failures(session, count=None, start=None, stop=None, sort_by=None, descending=None):
-    query = session.query(FailedEntry)
-    if count:
-        return query.count()
-    if descending:
-        query = query.order_by(getattr(FailedEntry, sort_by).desc())
-    else:
-        query = query.order_by(getattr(FailedEntry, sort_by))
-    return query.slice(start, stop).all()
